Rewrite for Elm 0.18
Basically the whole program had to be revamped due to architectural
changes in Elm 0.17 and 0.18. It's now based on
<https://github.com/lassik/elm-webmidi-ports-minimal>. `native.js` and
`src/WebMidi.elm` are copied verbatim from there.

Thanks to newlandsvalley for great work on MIDI in Elm. `elm-comidi`
is great and `elm-webmidi-ports` was instrumental in understanding how
to make all of this work in Elm 0.18.